newsite/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.3
python-dotenv==1.0.1
Werkzeug==2.3.7
SQLAlchemy==1.4.51
email-validator==2.1.0.post1
psycopg2-binary==2.9.9 